+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `ContactsFeature::findContacts` parameter bag made optional
+
 ## [2.1.3] - 2019-04-11
 ### Fixed
 - `CreateContactBag::withPhone`